refactor: deletar a antiga table e renomear arquivos/pastas
