+++ src/App.tsx
@@ -2,9 +2,8 @@ import React from 'react';
 import { BrowserRouter, Routes, Route, Link } from 'react-router-dom'
 import './App.css';
 import NavBar from './components/navBar'
-import Table from './pages/Table'
 import Home from './pages/Home'
-import NewTable from './pages/NewTable'
+import Table from './pages/Table'
 
 function App() {
   return (
@@ -16,7 +15,6 @@ function App() {
         <Routes>
           <Route path="/" element={<Home/>}/>
           <Route path="/tabela" element={<Table/>}/>
-          <Route path="/tabelanova" element={<NewTable/>}/>
         </Routes>
       </div>
     </BrowserRouter>
