+++ src/components/table/TableRowEdit.tsx
@@ -1,48 +1,39 @@
-import React from "react";
-import { useState } from "react";
-import { Row, TableRowEditProps } from "../../types";
+import React from "react"
+import { useState } from "react"
+import { RowData, TableRowEditMethods, TableRowEditProps } from "../../typesTable"
+import themes from '../themes/Table.module.css'
+import checkIcon from '../../assets/checkIcon.svg'
 
 function TableRowEdit(props: TableRowEditProps) {
-    const row = props.row
-    const data = props.data
-    type tableCell = keyof typeof row
-    const keys: tableCell[] = Object.keys(row).filter((key): key is keyof Row => key in row && key !== "id")
-    const [editedData, setEditedData] = useState<Row>(row)
+    const rowData = props.rowData
+    const [editedRowData, setEditedRowData] = useState<RowData>(rowData)
 
-    function updateEditedData(e: React.ChangeEvent<HTMLInputElement>) {
-        if(e.target.placeholder === 'nome') {
-            setEditedData({...editedData, nome: e.target.value})
-        }
-        if(e.target.placeholder === 'idade') {
-            setEditedData({...editedData, idade: parseInt(e.target.value)})
-        }
-    }
-    function sendEditRow(e: React.MouseEvent<HTMLButtonElement, MouseEvent>) {
-        let auxData = [...data]
-        const id = parseInt(e.currentTarget.name)
-        auxData.forEach((row, index) => {
-            if(row['id'] === id) {
-                auxData[index] = editedData
+    function updateEditedRowData(e: React.ChangeEvent<HTMLInputElement>) {
+        props.keys.map(key => {
+            if(key === e.target.name) {
+                setEditedRowData({...editedRowData, [key]: e.target.value})
             }
         })
-        props.rowMethods.setData(auxData)
-        props.rowMethods.setEditDataId(0)
     }
 
     return (
-        <tr key={row.id}>
-            <td key={'id'}>{row['id']}</td>
+        <tr>
+            <td key="id">{rowData.id}</td>
             {
-                keys.map( key => <td key={key}><input type="text" placeholder={key} onChange={updateEditedData}/></td>)
+                props.keys.filter(key => key !== "id").map( key => (
+                    <td key={key}>
+                        <input type="text" name={key} placeholder={rowData[key]} onChange={updateEditedRowData}/>
+                    </td>
+                ))
             }
             <td>
-                <button name={row.id+""} onClick={sendEditRow}>Enviar</button>
+                <button onClick={() => props.methods.editRow(editedRowData)} className={themes.enviar}> <img src={checkIcon} alt="Enviar" /></button>
             </td>
             <td>
-                <button name={row.id+""} onClick={props.rowMethods.removeRow}>X</button>
+                <button onClick={() => props.methods.removeRow(rowData.id)} className={themes.remover}>X</button>
             </td>
         </tr>
     )
 }
 
-export default TableRowEdit;
+export default TableRowEdit
